hari ke 3 masuk de CRUD dokter

+++ database/migrations/2022_10_16_143457_create_pasiens_table.php
@@ -17,9 +17,9 @@ return new class extends Migration
             $table->id();
             $table->integer('user_id');
             $table->string('name');
-            $table->integer('ktp');
-            $table->integer('bpjs');
-            $table->integer('no_hp');
+            $table->bigInteger('ktp');
+            $table->bigInteger('bpjs');
+            $table->bigInteger('no_hp');
             $table->string('jenis_kelamin');
             $table->date('tanggal_lahir');
             $table->text('alamat');
