+++ database/migrations/2022_10_16_144244_create_layanans_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('layanans', function (Blueprint $table) {
             $table->id();
             $table->string('nama_layanan');
-            $table->integer('harga_layanan');
+            $table->bigInteger('harga_layanan');
             $table->timestamps();
         });
     }
